restructurized endpoint tests - one file per endpoint
