+++ tests/TeamUp.EndToEndTests/EndpointTests/Teams/DeleteTeamTests.cs
@@ -1,0 +1,137 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace TeamUp.EndToEndTests.EndpointTests.Teams;
+
+public sealed class DeleteTeamTests : BaseTeamTests
+{
+	public DeleteTeamTests(TeamApiWebApplicationFactory appFactory) : base(appFactory) { }
+
+	[Fact]
+	public async Task DeleteTeam_AsOwner_Should_DeleteTeamInDatabase()
+	{
+		//arrange
+		var owner = UserGenerator.ActivatedUser.Generate();
+		var members = UserGenerator.ActivatedUser.Generate(19);
+		var team = TeamGenerator.GenerateTeamWith(owner, members);
+
+		await UseDbContextAsync(dbContext =>
+		{
+			dbContext.Users.Add(owner);
+			dbContext.Users.AddRange(members);
+			dbContext.Teams.Add(team);
+			return dbContext.SaveChangesAsync();
+		});
+
+		Authenticate(owner);
+
+		//act
+		var response = await Client.DeleteAsync($"/api/v1/teams/{team.Id.Value}");
+
+		//assert
+		response.Should().Be200Ok();
+
+		await UseDbContextAsync(async dbContext =>
+		{
+			var deletedTeam = await dbContext.Teams.FindAsync(team.Id);
+			deletedTeam.Should().BeNull();
+
+			var members = await dbContext.Set<TeamMember>().ToListAsync();
+			members.Should().BeEmpty();
+
+			var eventTypes = await dbContext.Set<EventType>().ToListAsync();
+			eventTypes.Should().BeEmpty();
+
+			var events = await dbContext.Events.ToListAsync();
+			events.Should().BeEmpty();
+
+			var eventResponses = await dbContext.Set<EventResponse>().ToListAsync();
+			eventResponses.Should().BeEmpty();
+		});
+	}
+
+	[Theory]
+	[InlineData(TeamRole.Member)]
+	[InlineData(TeamRole.Coordinator)]
+	[InlineData(TeamRole.Admin)]
+	public async Task DeleteTeam_AsAdminOrLower_Should_ResultInForbidden(TeamRole teamRole)
+	{
+		//arrange
+		var owner = UserGenerator.ActivatedUser.Generate();
+		var initiatorUser = UserGenerator.ActivatedUser.Generate();
+		var members = UserGenerator.ActivatedUser.Generate(18);
+		var team = TeamGenerator.GenerateTeamWith(owner, members, (initiatorUser, teamRole));
+
+		await UseDbContextAsync(dbContext =>
+		{
+			dbContext.Users.AddRange([owner, initiatorUser]);
+			dbContext.Users.AddRange(members);
+			dbContext.Teams.Add(team);
+			return dbContext.SaveChangesAsync();
+		});
+
+		Authenticate(initiatorUser);
+
+		//act
+		var response = await Client.DeleteAsync($"/api/v1/teams/{team.Id.Value}");
+
+		//assert
+		response.Should().Be403Forbidden();
+
+		var problemDetails = await response.Content.ReadFromJsonAsync<ProblemDetails>();
+		problemDetails.ShouldContainError(TeamErrors.UnauthorizedToDeleteTeam);
+	}
+
+	[Fact]
+	public async Task DeleteTeam_ThatDoesNotExist_Should_ResultInNotFound()
+	{
+		//arrange
+		var user = UserGenerator.ActivatedUser.Generate();
+		var teamId = F.Random.Guid();
+
+		await UseDbContextAsync(dbContext =>
+		{
+			dbContext.Users.Add(user);
+			return dbContext.SaveChangesAsync();
+		});
+
+		Authenticate(user);
+
+		//act
+		var response = await Client.DeleteAsync($"/api/v1/teams/{teamId}");
+
+		//assert
+		response.Should().Be404NotFound();
+
+		var problemDetails = await response.Content.ReadFromJsonAsync<ProblemDetails>();
+		problemDetails.ShouldContainError(TeamErrors.TeamNotFound);
+	}
+
+	[Fact]
+	public async Task DeleteTeam_WhenNotMemberOfTeam_Should_ResultInForbidden()
+	{
+		//arrange
+		var owner = UserGenerator.ActivatedUser.Generate();
+		var initiatorUser = UserGenerator.ActivatedUser.Generate();
+		var members = UserGenerator.ActivatedUser.Generate(19);
+		var team = TeamGenerator.GenerateTeamWith(owner, members);
+
+		await UseDbContextAsync(dbContext =>
+		{
+			dbContext.Users.AddRange([owner, initiatorUser]);
+			dbContext.Users.AddRange(members);
+			dbContext.Teams.Add(team);
+			return dbContext.SaveChangesAsync();
+		});
+
+		Authenticate(initiatorUser);
+
+		//act
+		var response = await Client.DeleteAsync($"/api/v1/teams/{team.Id.Value}");
+
+		//assert
+		response.Should().Be403Forbidden();
+
+		var problemDetails = await response.Content.ReadFromJsonAsync<ProblemDetails>();
+		problemDetails.ShouldContainError(TeamErrors.NotMemberOfTeam);
+	}
+}
